Change precio column type from Float to Numeric

+++ models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Float
+from sqlalchemy import Column, Integer, String, Numeric
 from database import Base
 
 class Producto(Base):
@@ -6,5 +6,5 @@ class Producto(Base):
 
     id = Column(Integer, primary_key=True, index=True)
     nombre = Column(String, nullable=False)
-    precio = Column(Float, nullable=False)
+    precio = Column(Numeric, nullable=False)
     stock = Column(Integer, nullable=False)
